core: print symlink destinations
Wow. This required a big overhaul. Because the list of files we were
keeping was simply a list of strings, there was no way to store both the
source _and_ destination of the link, so I had to create a list that can
store both simple filenames and also symlinks.

Not sure if my attempt at a generic list is the most C-like approach,
but I like that it leaves it open for extension to do something
different for other kinds of entities aswell (like block, and character
devices, and sockets, etc.), so I'll leave it for now.

I also don't like the number of heap allocations we're now doing,
although it hasn't resulted in any noticeable slowness on my MacBook in
my totally unscientific measurements.